Fixed issues reported in testing

+++ src/routes/resources/resources.ctrl.js
@@ -1,4 +1,5 @@
 import { Types } from 'mongoose';
+import { ERROR_KEYS } from '../../constants';
 import {
   BookingResource,
   ResourceCollectionModel,
@@ -165,20 +166,31 @@ export class ResourceController {
     return await ResourceModel.bulkWrite(bookingResources);
   }
   static async unassignResources(payload) {
+    const resourceId = payload.resources.pop();
+    const resource = await ResourceModel.findById(resourceId);
+    if (!resource) throw ERROR_KEYS.RESOURCE_NOT_FOUND;
     const query = [];
     payload.bookings.forEach(a => {
-      payload.resources.forEach(r => {
-        query.push({
-          deleteOne: {
-            filter: {
-              tripId: Types.ObjectId(payload.tripId),
-              resourceId: Types.ObjectId(r),
-              bookingId: Types.ObjectId(a.bookingId),
-            },
+      query.push({
+        deleteOne: {
+          filter: {
+            tripId: Types.ObjectId(payload.tripId),
+            resourceId: resource._id,
+            bookingId: Types.ObjectId(a.bookingId),
           },
-        });
+        },
       });
     });
+    const bIds = payload.bookings.map(b => b.bookingId);
+    const rooms = {};
+    Object.entries(resource.rooms).forEach(([columnId, column]) => {
+      if (column.items.length > 0) {
+        column.items = column.items.filter(b => !bIds.includes(b.bookingId));
+        column.attendees = column.items.reduce((s, a) => s + a.attendees, 0);
+      } else column.attendees = 0;
+      rooms[columnId] = column;
+    });
+    await ResourceModel.updateOne({ _id: resource._id }, { rooms });
     return await BookingResource.bulkWrite(query);
     // return await ResourceController.setAteendeeCount(payload);
   }
